soc/cores/interconnect: Rely on WaitTimer's new automatic cast to int.

+++ litex/soc/cores/esc.py
@@ -141,15 +141,15 @@ class ESCDShot(LiteXModule):
         timings.compute()
 
         # Timers.
-        t0h_timer = WaitTimer(int(timings.t0h*sys_clk_freq))
-        t0l_timer = WaitTimer(int(timings.t0l*sys_clk_freq) - 1) # Compensate Xfer FSM latency.
+        t0h_timer = WaitTimer(timings.t0h*sys_clk_freq)
+        t0l_timer = WaitTimer(timings.t0l*sys_clk_freq - 1) # Compensate Xfer FSM latency.
         self.submodules += t0h_timer, t0l_timer
 
-        t1h_timer = WaitTimer(int(timings.t1h*sys_clk_freq))
-        t1l_timer = WaitTimer(int(timings.t1l*sys_clk_freq) - 1) # Compensate Xfer FSM latency.
+        t1h_timer = WaitTimer(timings.t1h*sys_clk_freq)
+        t1l_timer = WaitTimer(timings.t1l*sys_clk_freq - 1) # Compensate Xfer FSM latency.
         self.submodules += t1h_timer, t1l_timer
 
-        tgap_timer = WaitTimer(int(timings.tgap*sys_clk_freq))
+        tgap_timer = WaitTimer(timings.tgap*sys_clk_freq)
         self.submodules += tgap_timer
 
         # XFER FSM.
